Setting replica identity in Entity table SQL script
Setting replica identity for Entity table in Postgres to enable logging
of "before" field in Debezium message. This needs to be enabled in all
Scorpio services due to the presence of SQL script in all services, not
doing so gives checksum mismatch issue when other services start up.

Resolves: Issue IndustryFusion/DigitalTwin#31
Related: Issue IndustryFusion/DigitalTwin#18