Replace hardcoded passwords for Scorpio DB
Replace instances of hardcoded passwords for DB user 'ngb' in all
application-dist.yml for ScorpioBroker services with environment
variables. The DB user password is randomly generated by the Zalando
Postgres Operator and stored in a kubernetes secret. When deploying
Scorpio using Helm, the environment variables are initialised with this
secret.

Resolves: IndustryFusion/DigitalTwin#67
Related: IndustryFusion/DigitalTwin#26